+++ webgpu-react/src/shaders/main/structs.wgsl
@@ -19,7 +19,7 @@ struct CellData {
     height       : f32,       // 0..3
     heightNormal : vec3<f32>, // starts at 16, uses 16 bytes
     fAmount      : f32,       // at 28
-    randDir      : u32,
+    randDir      : f32,
     // _pad0        : vec3<f32>, // starts at 32, uses 16 bytes
 };
 // total size = 48 bytes, 12 floats
